Base Dockerfiles for Falaise CI and Production
Update CentOS and Ubuntu Dockerfiles to provide a base install
of the system and linuxbrew requirements for Falaise.

Change USER to linuxbrew and installation of brew to $HOME/.linuxbrew
for future compatibility with upstream linuxbrew.

Use base Ubuntu 16.04 image to create image including and install
of Falaise. Update docker README to describe use and build of
images.